renaming autoregressive and logistic regression hmm
