+++ dynamax/hidden_markov_model/__init__.py
@@ -1,12 +1,12 @@
 from dynamax.hidden_markov_model.models.abstractions import HMM, HMMEmissions, HMMInitialState, HMMTransitions, HMMParameterSet, HMMPropertySet
-from dynamax.hidden_markov_model.models.autoregressive_hmm import LinearAutoregressiveHMM
+from dynamax.hidden_markov_model.models.arhmm import LinearAutoregressiveHMM
 from dynamax.hidden_markov_model.models.bernoulli_hmm import BernoulliHMM
 from dynamax.hidden_markov_model.models.categorical_glm_hmm import CategoricalRegressionHMM
 from dynamax.hidden_markov_model.models.categorical_hmm import CategoricalHMM
 from dynamax.hidden_markov_model.models.gaussian_hmm import GaussianHMM, DiagonalGaussianHMM, SphericalGaussianHMM, SharedCovarianceGaussianHMM, LowRankGaussianHMM
 from dynamax.hidden_markov_model.models.gmm_hmm import GaussianMixtureHMM, DiagonalGaussianMixtureHMM
 from dynamax.hidden_markov_model.models.linreg_hmm import LinearRegressionHMM
-from dynamax.hidden_markov_model.models.logistic_regression_hmm import LogisticRegressionHMM
+from dynamax.hidden_markov_model.models.logreg_hmm import LogisticRegressionHMM
 from dynamax.hidden_markov_model.models.multinomial_hmm import MultinomialHMM
 from dynamax.hidden_markov_model.models.poisson_hmm import PoissonHMM
