+++ dynamax/hidden_markov_model/models/arhmm.py
